solve cygwin compilation errors and warnings

+++ src/pescan.c
@@ -249,7 +249,7 @@ bool strisprint(const char *string)
 
 	while (*s)
 	{
-		if (!isalpha(*s))
+		if (!isalpha((int)*s))
 			return false;
 
 		s++;
